+++ src/templates/foot.js
@@ -1,7 +1,0 @@
-const compile = require('string-template/compile');
-
-const template = compile(`
-  </mj-body>
-</mjml>`);
-
-module.exports = template;