+++ whatnode.js
@@ -1,1 +1,0 @@
-module.exports = cb => cb(null, process.versions)